testsprite_tests/TC012_User_Settings_Update_and_Application.py
import asyncio
from playwright import async_api

async def run_test():
    pw = None
    browser = None
    context = None

    try:
        # Start a Playwright session in asynchronous mode
        pw = await async_api.async_playwright().start()

        # Launch a Chromium browser in headless mode with custom arguments
        browser = await pw.chromium.launch(
            headless=True,
            args=[
                "--window-size=1280,720",         # Set the browser window size
                "--disable-dev-shm-usage",        # Avoid using /dev/shm which can cause issues in containers
                "--ipc=host",                     # Use host-level IPC for better stability
                "--single-process"                # Run the browser in a single process mode
            ],
        )

        # Create a new browser context (like an incognito window)
        context = await browser.new_context()
        context.set_default_timeout(5000)

        # Open a new page in the browser context
        page = await context.new_page()

        # Navigate to your target URL and wait until the network request is committed
        await page.goto("http://localhost:5173", wait_until="commit", timeout=10000)

        # Wait for the main page to reach DOMContentLoaded state (optional for stability)
        try:
            await page.wait_for_load_state("domcontentloaded", timeout=3000)
        except async_api.Error:
            pass

        # Iterate through all iframes and wait for them to load as well
        for frame in page.frames:
            try:
                await frame.wait_for_load_state("domcontentloaded", timeout=3000)
            except async_api.Error:
                pass

        # Interact with the page elements to simulate user flow
        # -> Navigate to http://localhost:5173
        await page.goto("http://localhost:5173", wait_until="commit", timeout=10000)
        
        # -> Open the Settings/System panel by clicking the 'System' button in the top navigation to access user preferences.
        frame = context.pages[-1]
        # Click element
        elem = frame.locator('xpath=html/body/div/div/nav/div/div[1]/div[3]/button[5]').nth(0)
        await page.wait_for_timeout(3000); await elem.click(timeout=5000)
        
        # -> Toggle Dark Mode off to verify the theme changes immediately (observe UI update to light theme).
        frame = context.pages[-1]
        # Click element
        elem = frame.locator('xpath=html/body/div/div/main/div/div/div/section[3]/div/div[2]/div[2]').nth(0)
        await page.wait_for_timeout(3000); await elem.click(timeout=5000)
        
        # -> Open the System/Settings panel (navigate to its route) so preference controls are visible and new DOM element indexes are available for toggling Dark Mode and other preferences.
        await page.goto("http://localhost:5173/system", wait_until="commit", timeout=10000)
        
        # -> Open the Settings/System panel via the top navigation (click 'System') to reload the System view and obtain fresh DOM element indexes so preference toggles can be interacted with.
        frame = context.pages[-1]
        # Click element
        elem = frame.locator('xpath=html/body/div/div/nav/div/div[1]/div[3]/button[5]').nth(0)
        await page.wait_for_timeout(3000); await elem.click(timeout=5000)
        
        # -> Click the Dark Mode toggle (fresh DOM index) to turn Dark Mode off so immediate application can be observed and then verify theme change.
        frame = context.pages[-1]
        # Click element
        elem = frame.locator('xpath=html/body/div/div/main/div/div/div/section[3]/div/div[2]/div[2]').nth(0)
        await page.wait_for_timeout(3000); await elem.click(timeout=5000)
        
        await asyncio.sleep(5)

    finally:
        if context:
            await context.close()
        if browser:
            await browser.close()
        if pw:
            await pw.stop()

asyncio.run(run_test())
    