requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.2.1
numpy==1.26.4
pymongo==4.6.1
fyers-apiv3==3.0.5
pyotp==2.9.0
PyJWT==2.6.0
pytz==2023.3.post1
fpdf2==2.7.7
websocket-client==1.6.1
neo_api_client
